requirements.txt
imageio
imageio-ffmpeg
importlib-metadata>=1.5.0
ipykernel>=5.1.1
IPython>=7.7.0
magicgui==0.2.8
napari-plugin-engine>=0.1.5
napari-svg>=0.1.3
natsort
numpy>=1.10
pandas
Pillow!=7.1.0,!=7.1.1  # not a direct dependency, but 7.1.0 and 7.1.1 broke imageio
psutil>=5.0
PyOpenGL>=3.1.0
pyside2>=5.12.3, <5.15
PyYAML>=5.1
qtconsole>=4.5.1
qtpy>=1.7.0
scipy>=1.2.0
scikit-image==0.17.2
toolz>=0.10.0
vispy>=0.6.4
wrapt>=1.11.1
